api/requirements.txt
flask
flask-restx
flask-mongoengine
python-dotenv